[REFACTOR] - Separate search logic and UI into custom hook and component
